backend/requirements.txt
# c:\Obsidian\Ethyx\backend\requirements.txt
# Python 3.10.11 EXACT — do not upgrade Python version
# ════════════════════════════════════════════════════

# ─── Web Framework ─────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ─── Auth ──────────────────────────────────────────
PyJWT==2.8.0
cryptography==43.0.0

# ─── Database ──────────────────────────────────────
asyncpg==0.29.0
supabase==2.6.0

# ─── Azure Storage ─────────────────────────────────
azure-storage-blob==12.19.0
azure-identity==1.17.1

# ─── Gemini AI ─────────────────────────────────────
google-generativeai==0.8.3

# ─── Fairness / ML ─────────────────────────────────
aif360==0.6.1
fairlearn==0.10.0
pandas==2.2.2
scikit-learn==1.5.1

# CRITICAL: numpy must stay < 2.0 — AIF360 breaks on numpy 2.x
numpy>=1.23.0,<2.0.0

# ─── Rate Limiting ─────────────────────────────────
slowapi==0.1.9

# ─── Observability ─────────────────────────────────
sentry-sdk[fastapi]==2.13.0

# ─── Utilities ─────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0
